+++ lib/bundler/rubygems_integration.rb
@@ -93,6 +93,10 @@ module Bundler
       Gem.path
     end
 
+    def gem_cache
+      gem_path.map{|p| File.expand_path("cache", p) }
+    end
+
     def spec_cache_dirs
       @spec_cache_dirs ||= begin
         dirs = gem_path.map {|dir| File.join(dir, 'specifications')}