+++ spec/bundler/cli_spec.rb
@@ -1,6 +1,9 @@
 require 'spec_helper'
+require 'bundler/cli'
 
 describe "bundle executable" do
+  let(:source_uri) { "http://localgemserver.test" }
+
   it "returns non-zero exit status when passed unrecognized options" do
     bundle '--invalid_argument', :exitstatus => true
     expect(exitstatus).to_not be_zero