+++ spec/spec_helper.rb
@@ -63,6 +63,14 @@ RSpec.configure do |config|
     config.filter_run_excluding :ruby => "1.9"
   end
 
+  if RUBY_VERSION >= "2.0"
+    config.filter_run_excluding :ruby => "1.8"
+    config.filter_run_excluding :ruby => "1.9"
+  else
+    config.filter_run_excluding :ruby => "2.0"
+    config.filter_run_excluding :ruby => "2.1"
+  end
+
   if Gem::VERSION < "2.2"
     config.filter_run_excluding :rubygems => "2.2"
   end